plots/plot_requirements.txt
awkward==1.10.5
awkward0==0.15.5
boost-histogram==1.4.0
cachetools==5.3.2
click==8.1.7
cloudpickle==3.0.0
coffea==0.7.22
contourpy==1.1.0
correctionlib==2.4.0
cycler==0.11.0
fonttools==4.40.0
fsspec==2023.10.0
hist==2.7.2
histoprint==2.4.0
importlib-resources==5.12.0
kiwisolver==1.4.4
llvmlite==0.41.1
lz4==4.3.2
markdown-it-py==3.0.0
matplotlib==3.7.1
mdurl==0.1.2
mplhep==0.3.31
mplhep-data==0.0.3
numba==0.58.1
numpy==1.23.5
packaging==23.1
pandas==2.1.3
pep8==1.7.1
Pillow==9.5.0
pyarrow==14.0.1
pycodestyle==2.11.1
pydantic==1.10.13
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rich==13.7.0
scipy==1.11.4
six==1.16.0
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uhi==0.4.0
uproot==4.3.7
uproot3==3.14.4
uproot3-methods==0.10.1
zipp==3.15.0
